requirements.txt
streamlit>=1.31.0
pydantic>=2.6.0
google-generativeai>=0.3.0
wikipedia>=1.4.0
python-dotenv>=1.0.1
